Move `StacksTransaction`
