+++ stacks-codec/src/transaction.rs
@@ -22,6 +22,7 @@
 
 use std::error;
 use std::fmt::{self, Display};
+use std::hash::Hash;
 use std::io::{Read, Write};
 
 use clarity_types::representations::{ClarityName, ContractName};
@@ -42,7 +43,7 @@ use stacks_common::types::chainstate::{
     StacksPublicKey, Txid,
 };
 use stacks_common::types::{PrivateKey, StacksEpochId, StacksPublicKeyBuffer};
-use stacks_common::util::hash::{Hash160, Sha512Trunc256Sum};
+use stacks_common::util::hash::{Hash160, MerkleHashFunc, MerkleTree, Sha512Trunc256Sum};
 use stacks_common::util::retry::BoundReader;
 use stacks_common::util::secp256k1::{MessageSignature, MESSAGE_SIGNATURE_ENCODED_SIZE};
 use stacks_common::util::vrf::VRFProof;
@@ -2817,5 +2818,664 @@ impl From<TransactionSmartContract> for TransactionPayload {
 impl From<TransactionContractCall> for TransactionPayload {
     fn from(value: TransactionContractCall) -> Self {
         TransactionPayload::ContractCall(value)
+    }
+}
+
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct StacksTransaction {
+    pub version: TransactionVersion,
+    pub chain_id: u32,
+    pub auth: TransactionAuth,
+    pub anchor_mode: TransactionAnchorMode,
+    pub post_condition_mode: TransactionPostConditionMode,
+    pub post_conditions: Vec<TransactionPostCondition>,
+    pub payload: TransactionPayload,
+}
+
+#[derive(Debug, Clone, PartialEq)]
+pub struct StacksTransactionSigner {
+    pub tx: StacksTransaction,
+    pub sighash: Txid,
+    origin_done: bool,
+    check_oversign: bool,
+    check_overlap: bool,
+}
+
+impl<'a, H> FromIterator<&'a StacksTransaction> for MerkleTree<H>
+where
+    H: MerkleHashFunc + Clone + PartialEq + fmt::Debug,
+{
+    fn from_iter<T: IntoIterator<Item = &'a StacksTransaction>>(iter: T) -> Self {
+        let txid_vec: Vec<_> = iter
+            .into_iter()
+            .map(|x| x.txid().as_bytes().to_vec())
+            .collect();
+        MerkleTree::new(&txid_vec)
+    }
+}
+
+impl Hash for StacksTransaction {
+    fn hash<H: std::hash::Hasher>(&self, state: &mut H) {
+        self.txid().hash(state);
+    }
+}
+
+impl Eq for StacksTransaction {}
+
+impl StacksMessageCodec for StacksTransaction {
+    fn consensus_serialize<W: Write>(&self, fd: &mut W) -> Result<(), codec_error> {
+        write_next(fd, &(self.version as u8))?;
+        write_next(fd, &self.chain_id)?;
+        write_next(fd, &self.auth)?;
+        write_next(fd, &(self.anchor_mode as u8))?;
+        write_next(fd, &(self.post_condition_mode as u8))?;
+        write_next(fd, &self.post_conditions)?;
+        write_next(fd, &self.payload)?;
+        Ok(())
+    }
+
+    fn consensus_deserialize<R: Read>(fd: &mut R) -> Result<StacksTransaction, codec_error> {
+        StacksTransaction::consensus_deserialize_with_len(fd).map(|(result, _)| result)
+    }
+}
+
+impl StacksTransaction {
+    pub fn tx_len(&self) -> u64 {
+        let mut tx_bytes = vec![];
+        self.consensus_serialize(&mut tx_bytes)
+            .expect("BUG: Failed to serialize a transaction object");
+        u64::try_from(tx_bytes.len()).expect("tx len exceeds 2^64 bytes")
+    }
+
+    pub fn consensus_deserialize_with_len<R: Read>(
+        fd: &mut R,
+    ) -> Result<(StacksTransaction, u64), codec_error> {
+        let mut bound_read = BoundReader::from_reader(fd, MAX_TRANSACTION_LEN.into());
+        let fd = &mut bound_read;
+
+        let version_u8: u8 = read_next(fd)?;
+        let chain_id: u32 = read_next(fd)?;
+        let auth: TransactionAuth = read_next(fd)?;
+        let anchor_mode_u8: u8 = read_next(fd)?;
+        let post_condition_mode_u8: u8 = read_next(fd)?;
+        let post_conditions: Vec<TransactionPostCondition> = read_next(fd)?;
+
+        let payload: TransactionPayload = read_next(fd)?;
+
+        let version = if (version_u8 & 0x80) == 0 {
+            TransactionVersion::Mainnet
+        } else {
+            TransactionVersion::Testnet
+        };
+
+        let anchor_mode = match anchor_mode_u8 {
+            x if x == TransactionAnchorMode::OffChainOnly as u8 => {
+                TransactionAnchorMode::OffChainOnly
+            }
+            x if x == TransactionAnchorMode::OnChainOnly as u8 => {
+                TransactionAnchorMode::OnChainOnly
+            }
+            x if x == TransactionAnchorMode::Any as u8 => TransactionAnchorMode::Any,
+            _ => {
+                return Err(codec_error::DeserializeError(format!(
+                    "Failed to parse transaction: invalid anchor mode {}",
+                    anchor_mode_u8
+                )));
+            }
+        };
+
+        // if the payload is a proof of a poisoned microblock stream, or is a coinbase, then this _must_ be anchored.
+        // Otherwise, if the offending leader is the next leader, they can just orphan their proof
+        // of malfeasance.
+        match payload {
+            TransactionPayload::PoisonMicroblock(_, _) => {
+                if anchor_mode != TransactionAnchorMode::OnChainOnly {
+                    return Err(codec_error::DeserializeError(
+                        "Failed to parse transaction: invalid anchor mode for PoisonMicroblock"
+                            .to_string(),
+                    ));
+                }
+            }
+            TransactionPayload::Coinbase(..) => {
+                if anchor_mode != TransactionAnchorMode::OnChainOnly {
+                    return Err(codec_error::DeserializeError(
+                        "Failed to parse transaction: invalid anchor mode for Coinbase".to_string(),
+                    ));
+                }
+            }
+            _ => {}
+        }
+
+        let post_condition_mode = match post_condition_mode_u8 {
+            x if x == TransactionPostConditionMode::Allow as u8 => {
+                TransactionPostConditionMode::Allow
+            }
+            x if x == TransactionPostConditionMode::Deny as u8 => {
+                TransactionPostConditionMode::Deny
+            }
+            x if x == TransactionPostConditionMode::Originator as u8 => {
+                TransactionPostConditionMode::Originator
+            }
+            _ => {
+                return Err(codec_error::DeserializeError(format!(
+                    "Failed to parse transaction: invalid post-condition mode {}",
+                    post_condition_mode_u8
+                )));
+            }
+        };
+        let tx = StacksTransaction {
+            version,
+            chain_id,
+            auth,
+            anchor_mode,
+            post_condition_mode,
+            post_conditions,
+            payload,
+        };
+
+        Ok((tx, fd.num_read()))
+    }
+
+    /// Try to convert to a coinbase payload
+    pub fn try_as_coinbase(
+        &self,
+    ) -> Option<(&CoinbasePayload, Option<&PrincipalData>, Option<&VRFProof>)> {
+        match &self.payload {
+            TransactionPayload::Coinbase(payload, recipient_opt, vrf_proof_opt) => {
+                Some((payload, recipient_opt.as_ref(), vrf_proof_opt.as_ref()))
+            }
+            _ => None,
+        }
+    }
+
+    /// Try to convert to a tenure change payload
+    pub fn try_as_tenure_change(&self) -> Option<&TenureChangePayload> {
+        match &self.payload {
+            TransactionPayload::TenureChange(tc_payload) => Some(tc_payload),
+            _ => None,
+        }
+    }
+
+    /// Create a new, unsigned transaction and an empty STX fee with no post-conditions.
+    pub fn new(
+        version: TransactionVersion,
+        auth: TransactionAuth,
+        payload: TransactionPayload,
+    ) -> StacksTransaction {
+        let anchor_mode = match payload {
+            TransactionPayload::Coinbase(..) => TransactionAnchorMode::OnChainOnly,
+            TransactionPayload::PoisonMicroblock(_, _) => TransactionAnchorMode::OnChainOnly,
+            _ => TransactionAnchorMode::Any,
+        };
+
+        StacksTransaction {
+            version,
+            chain_id: 0,
+            auth,
+            anchor_mode,
+            post_condition_mode: TransactionPostConditionMode::Deny,
+            post_conditions: vec![],
+            payload,
+        }
+    }
+
+    /// Get fee rate
+    pub fn get_tx_fee(&self) -> u64 {
+        self.auth.get_tx_fee()
+    }
+
+    /// Set fee rate
+    pub fn set_tx_fee(&mut self, tx_fee: u64) {
+        self.auth.set_tx_fee(tx_fee);
+    }
+
+    /// Get origin nonce
+    pub fn get_origin_nonce(&self) -> u64 {
+        self.auth.get_origin_nonce()
+    }
+
+    /// get sponsor nonce
+    pub fn get_sponsor_nonce(&self) -> Option<u64> {
+        self.auth.get_sponsor_nonce()
+    }
+
+    /// set origin nonce
+    pub fn set_origin_nonce(&mut self, n: u64) {
+        self.auth.set_origin_nonce(n);
+    }
+
+    /// set sponsor nonce
+    pub fn set_sponsor_nonce(&mut self, n: u64) -> Result<(), AuthError> {
+        self.auth.set_sponsor_nonce(n)
+    }
+
+    /// Set anchor mode
+    pub fn set_anchor_mode(&mut self, anchor_mode: TransactionAnchorMode) {
+        self.anchor_mode = anchor_mode;
+    }
+
+    /// Set post-condition mode
+    pub fn set_post_condition_mode(&mut self, postcond_mode: TransactionPostConditionMode) {
+        self.post_condition_mode = postcond_mode;
+    }
+
+    /// Add a post-condition
+    pub fn add_post_condition(&mut self, post_condition: TransactionPostCondition) {
+        self.post_conditions.push(post_condition);
+    }
+
+    /// a txid of a stacks transaction is its sha512/256 hash
+    pub fn txid(&self) -> Txid {
+        let mut bytes = vec![];
+        self.consensus_serialize(&mut bytes)
+            .expect("BUG: failed to serialize to a vec");
+        Txid::from_stacks_tx(&bytes)
+    }
+
+    /// Get a mutable reference to the internal auth structure
+    pub fn borrow_auth(&mut self) -> &mut TransactionAuth {
+        &mut self.auth
+    }
+
+    /// Get an immutable reference to the internal auth structure
+    pub fn auth(&self) -> &TransactionAuth {
+        &self.auth
+    }
+
+    /// begin signing the transaction.
+    /// If this is a sponsored transaction, then the origin only commits to knowing that it is
+    /// sponsored.  It does _not_ commit to the sponsored fields, so set them all to sentinel
+    /// values.
+    /// Return the initial sighash.
+    pub fn sign_begin(&self) -> Txid {
+        let mut tx = self.clone();
+        tx.auth = tx.auth.into_initial_sighash_auth();
+        tx.txid()
+    }
+
+    /// begin verifying a transaction.
+    /// return the initial sighash
+    pub fn verify_begin(&self) -> Txid {
+        let mut tx = self.clone();
+        tx.auth = tx.auth.into_initial_sighash_auth();
+        tx.txid()
+    }
+
+    /// Sign a sighash and append the signature and public key to the given spending condition.
+    /// Returns the next sighash
+    fn sign_and_append(
+        condition: &mut TransactionSpendingCondition,
+        cur_sighash: &Txid,
+        auth_flag: &TransactionAuthFlags,
+        privk: &StacksPrivateKey,
+    ) -> Result<Txid, AuthError> {
+        let (next_sig, next_sighash) = TransactionSpendingCondition::next_signature(
+            cur_sighash,
+            auth_flag,
+            condition.tx_fee(),
+            condition.nonce(),
+            privk,
+        )?;
+        match condition {
+            TransactionSpendingCondition::Singlesig(ref mut cond) => {
+                cond.set_signature(next_sig);
+                Ok(next_sighash)
+            }
+            TransactionSpendingCondition::Multisig(ref mut cond) => {
+                cond.push_signature(
+                    if privk.compress_public() {
+                        TransactionPublicKeyEncoding::Compressed
+                    } else {
+                        TransactionPublicKeyEncoding::Uncompressed
+                    },
+                    next_sig,
+                );
+                Ok(next_sighash)
+            }
+            TransactionSpendingCondition::OrderIndependentMultisig(ref mut cond) => {
+                cond.push_signature(
+                    if privk.compress_public() {
+                        TransactionPublicKeyEncoding::Compressed
+                    } else {
+                        TransactionPublicKeyEncoding::Uncompressed
+                    },
+                    next_sig,
+                );
+                Ok(cur_sighash.clone())
+            }
+        }
+    }
+
+    /// Pop the last auth field
+    fn pop_auth_field(
+        condition: &mut TransactionSpendingCondition,
+    ) -> Option<TransactionAuthField> {
+        match condition {
+            TransactionSpendingCondition::Multisig(ref mut cond) => cond.pop_auth_field(),
+            TransactionSpendingCondition::OrderIndependentMultisig(ref mut cond) => {
+                cond.pop_auth_field()
+            }
+            TransactionSpendingCondition::Singlesig(ref mut cond) => cond.pop_signature(),
+        }
+    }
+
+    /// Append a public key to a multisig condition
+    fn append_pubkey(
+        condition: &mut TransactionSpendingCondition,
+        pubkey: &StacksPublicKey,
+    ) -> Result<(), AuthError> {
+        match condition {
+            TransactionSpendingCondition::Multisig(ref mut cond) => {
+                cond.push_public_key(pubkey.clone());
+                Ok(())
+            }
+            TransactionSpendingCondition::OrderIndependentMultisig(ref mut cond) => {
+                cond.push_public_key(pubkey.clone());
+                Ok(())
+            }
+            _ => Err(AuthError::SigningError(
+                "Not a multisig condition".to_string(),
+            )),
+        }
+    }
+
+    /// Append the next signature from the origin account authorization.
+    /// Return the next sighash.
+    pub fn sign_next_origin(
+        &mut self,
+        cur_sighash: &Txid,
+        privk: &StacksPrivateKey,
+    ) -> Result<Txid, AuthError> {
+        let next_sighash = match self.auth {
+            TransactionAuth::Standard(ref mut origin_condition)
+            | TransactionAuth::Sponsored(ref mut origin_condition, _) => {
+                StacksTransaction::sign_and_append(
+                    origin_condition,
+                    cur_sighash,
+                    &TransactionAuthFlags::AuthStandard,
+                    privk,
+                )?
+            }
+        };
+        Ok(next_sighash)
+    }
+
+    /// Append the next public key to the origin account authorization.
+    pub fn append_next_origin(&mut self, pubk: &StacksPublicKey) -> Result<(), AuthError> {
+        match self.auth {
+            TransactionAuth::Standard(ref mut origin_condition) => {
+                StacksTransaction::append_pubkey(origin_condition, pubk)
+            }
+            TransactionAuth::Sponsored(ref mut origin_condition, _) => {
+                StacksTransaction::append_pubkey(origin_condition, pubk)
+            }
+        }
+    }
+
+    /// Append the next signature from the sponsoring account.
+    /// Return the next sighash
+    pub fn sign_next_sponsor(
+        &mut self,
+        cur_sighash: &Txid,
+        privk: &StacksPrivateKey,
+    ) -> Result<Txid, AuthError> {
+        let next_sighash = match self.auth {
+            TransactionAuth::Standard(_) => {
+                // invalid
+                return Err(AuthError::SigningError(
+                    "Cannot sign standard authorization with a sponsoring private key".to_string(),
+                ));
+            }
+            TransactionAuth::Sponsored(_, ref mut sponsor_condition) => {
+                StacksTransaction::sign_and_append(
+                    sponsor_condition,
+                    cur_sighash,
+                    &TransactionAuthFlags::AuthSponsored,
+                    privk,
+                )?
+            }
+        };
+        Ok(next_sighash)
+    }
+
+    /// Append the next public key to the sponsor account authorization.
+    pub fn append_next_sponsor(&mut self, pubk: &StacksPublicKey) -> Result<(), AuthError> {
+        match self.auth {
+            TransactionAuth::Standard(_) => Err(AuthError::SigningError(
+                "Cannot appned a public key to the sponsor of a standard auth condition"
+                    .to_string(),
+            )),
+            TransactionAuth::Sponsored(_, ref mut sponsor_condition) => {
+                StacksTransaction::append_pubkey(sponsor_condition, pubk)
+            }
+        }
+    }
+
+    /// Verify this transaction's signatures
+    pub fn verify(&self) -> Result<(), AuthError> {
+        self.auth.verify(&self.verify_begin())
+    }
+
+    /// Verify the transaction's origin signatures only.
+    /// Used by sponsors to get the next sig-hash to sign.
+    pub fn verify_origin(&self) -> Result<Txid, AuthError> {
+        self.auth.verify_origin(&self.verify_begin())
+    }
+
+    /// Get the origin account's address
+    pub fn origin_address(&self) -> StacksAddress {
+        match (&self.version, &self.auth) {
+            (TransactionVersion::Mainnet, TransactionAuth::Standard(origin_condition)) => {
+                origin_condition.address_mainnet()
+            }
+            (TransactionVersion::Testnet, TransactionAuth::Standard(origin_condition)) => {
+                origin_condition.address_testnet()
+            }
+            (
+                TransactionVersion::Mainnet,
+                TransactionAuth::Sponsored(origin_condition, _unused),
+            ) => origin_condition.address_mainnet(),
+            (
+                TransactionVersion::Testnet,
+                TransactionAuth::Sponsored(origin_condition, _unused),
+            ) => origin_condition.address_testnet(),
+        }
+    }
+
+    /// Get the sponsor account's address, if this transaction is sponsored
+    pub fn sponsor_address(&self) -> Option<StacksAddress> {
+        match (&self.version, &self.auth) {
+            (TransactionVersion::Mainnet, TransactionAuth::Standard(_unused)) => None,
+            (TransactionVersion::Testnet, TransactionAuth::Standard(_unused)) => None,
+            (
+                TransactionVersion::Mainnet,
+                TransactionAuth::Sponsored(_unused, sponsor_condition),
+            ) => Some(sponsor_condition.address_mainnet()),
+            (
+                TransactionVersion::Testnet,
+                TransactionAuth::Sponsored(_unused, sponsor_condition),
+            ) => Some(sponsor_condition.address_testnet()),
+        }
+    }
+
+    /// Get a copy of the origin spending condition
+    pub fn get_origin(&self) -> TransactionSpendingCondition {
+        self.auth.origin().clone()
+    }
+
+    /// Get a copy of the sending condition that will pay the tx fee
+    pub fn get_payer(&self) -> TransactionSpendingCondition {
+        match self.auth.sponsor() {
+            Some(tsc) => tsc.clone(),
+            None => self.auth.origin().clone(),
+        }
+    }
+
+    /// Is this a mainnet transaction?  false means 'testnet'
+    pub fn is_mainnet(&self) -> bool {
+        self.version == TransactionVersion::Mainnet
+    }
+
+    /// Is this a phantom transaction?
+    pub fn is_phantom(&self) -> bool {
+        let boot_address = StacksAddress::burn_address(self.is_mainnet()).into();
+        if let TransactionPayload::TokenTransfer(address, amount, _) = &self.payload {
+            *address == boot_address && *amount == 0
+        } else {
+            false
+        }
+    }
+}
+
+impl StacksTransactionSigner {
+    pub fn new(tx: &StacksTransaction) -> StacksTransactionSigner {
+        StacksTransactionSigner {
+            tx: tx.clone(),
+            sighash: tx.sign_begin(),
+            origin_done: false,
+            check_oversign: true,
+            check_overlap: true,
+        }
+    }
+
+    pub fn new_sponsor(
+        tx: &StacksTransaction,
+        spending_condition: TransactionSpendingCondition,
+    ) -> Result<StacksTransactionSigner, AuthError> {
+        if !tx.auth.is_sponsored() {
+            return Err(AuthError::IncompatibleSpendingConditionError);
+        }
+        let mut new_tx = tx.clone();
+        new_tx.auth.set_sponsor(spending_condition)?;
+        let origin_sighash = new_tx.verify_origin()?;
+
+        Ok(StacksTransactionSigner {
+            tx: new_tx,
+            sighash: origin_sighash,
+            origin_done: true,
+            check_oversign: true,
+            check_overlap: true,
+        })
+    }
+
+    pub fn resume(&mut self, tx: &StacksTransaction) {
+        self.tx = tx.clone()
+    }
+
+    pub fn disable_checks(&mut self) {
+        self.check_oversign = false;
+        self.check_overlap = false;
+    }
+
+    pub fn sign_origin(&mut self, privk: &StacksPrivateKey) -> Result<(), AuthError> {
+        if self.check_overlap && self.origin_done {
+            // can't sign another origin private key since we started signing sponsors
+            return Err(AuthError::SigningError(
+                "Cannot sign origin after sponsor key".to_string(),
+            ));
+        }
+
+        match self.tx.auth {
+            TransactionAuth::Standard(ref origin_condition) => {
+                if self.check_oversign
+                    && origin_condition.num_signatures() >= origin_condition.signatures_required()
+                {
+                    return Err(AuthError::SigningError(
+                        "Origin would have too many signatures".to_string(),
+                    ));
+                }
+            }
+            TransactionAuth::Sponsored(ref origin_condition, _) => {
+                if self.check_oversign
+                    && origin_condition.num_signatures() >= origin_condition.signatures_required()
+                {
+                    return Err(AuthError::SigningError(
+                        "Origin would have too many signatures".to_string(),
+                    ));
+                }
+            }
+        }
+
+        let next_sighash = self.tx.sign_next_origin(&self.sighash, privk)?;
+        self.sighash = next_sighash;
+        Ok(())
+    }
+
+    pub fn append_origin(&mut self, pubk: &StacksPublicKey) -> Result<(), AuthError> {
+        if self.check_overlap && self.origin_done {
+            // can't append another origin key
+            return Err(AuthError::SigningError(
+                "Cannot append public key to origin after sponsor key".to_string(),
+            ));
+        }
+
+        self.tx.append_next_origin(pubk)
+    }
+
+    pub fn sign_sponsor(&mut self, privk: &StacksPrivateKey) -> Result<(), AuthError> {
+        if let TransactionAuth::Sponsored(_, ref sponsor_condition) = self.tx.auth {
+            if self.check_oversign
+                && sponsor_condition.num_signatures() >= sponsor_condition.signatures_required()
+            {
+                return Err(AuthError::SigningError(
+                    "Sponsor would have too many signatures".to_string(),
+                ));
+            }
+        }
+
+        let next_sighash = self.tx.sign_next_sponsor(&self.sighash, privk)?;
+        self.sighash = next_sighash;
+        self.origin_done = true;
+        Ok(())
+    }
+
+    pub fn append_sponsor(&mut self, pubk: &StacksPublicKey) -> Result<(), AuthError> {
+        self.tx.append_next_sponsor(pubk)
+    }
+
+    pub fn pop_origin_auth_field(&mut self) -> Option<TransactionAuthField> {
+        match self.tx.auth {
+            TransactionAuth::Standard(ref mut origin_condition) => {
+                StacksTransaction::pop_auth_field(origin_condition)
+            }
+            TransactionAuth::Sponsored(ref mut origin_condition, _) => {
+                StacksTransaction::pop_auth_field(origin_condition)
+            }
+        }
+    }
+
+    pub fn pop_sponsor_auth_field(&mut self) -> Option<TransactionAuthField> {
+        match self.tx.auth {
+            TransactionAuth::Sponsored(_, ref mut sponsor_condition) => {
+                StacksTransaction::pop_auth_field(sponsor_condition)
+            }
+            _ => None,
+        }
+    }
+
+    pub fn complete(&self) -> bool {
+        match self.tx.auth {
+            TransactionAuth::Standard(ref origin_condition) => {
+                origin_condition.num_signatures() >= origin_condition.signatures_required()
+            }
+            TransactionAuth::Sponsored(ref origin_condition, ref sponsored_condition) => {
+                origin_condition.num_signatures() >= origin_condition.signatures_required()
+                    && sponsored_condition.num_signatures()
+                        >= sponsored_condition.signatures_required()
+                    && (self.origin_done || !self.check_overlap)
+            }
+        }
+    }
+
+    pub fn get_tx_incomplete(&self) -> StacksTransaction {
+        self.tx.clone()
+    }
+
+    pub fn get_tx(&self) -> Option<StacksTransaction> {
+        if self.complete() {
+            Some(self.tx.clone())
+        } else {
+            None
+        }
     }
 }
